requirements.txt
requests~=2.31.0
beautifulsoup4~=4.12
pandas~=2.2
sqlalchemy~=2.0
psycopg2-binary~=2.9
google-api-python-client~=2.120
google-auth~=2.29
google-auth-oauthlib~=1.2
pytest